refactor: inline IOS session table parsing
